[#47]feat: 마이페이지 통합
가이드와 여행자 모두가 공통으로 보는 화면인 마이페이지를 통합합니다.
+ 공지 발송 프로토타입 페이지 정리
